requirements.txt
# Core dependencies
torch>=1.12.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
PyYAML>=6.0
omegaconf>=2.2.0
tqdm>=4.64.0
pathlib2>=2.3.0
safetensors
tensordict
huggingface_hub
genqc

# Machine Learning
scikit-learn>=1.0.0

# Quantum Computing
qiskit>=0.39.0
qiskit-aer>=0.11.0

# Optional: genQC (needs to be installed separately)
# See: https://github.com/FlorianFuerrutter/genQC

# Development and testing (install with pip install -e .[dev])
# pytest>=7.0
# pytest-cov
# black
# flake8
# mypy
# jupyter
# notebook

# Documentation (install with pip install -e .[docs])  
# sphinx
# sphinx-rtd-theme
# myst-parser